UUIDField is only available in Django>=1.8
